// src/types/next_event.d.ts
export interface NextEvent {
	RoundNumber: number
	Country: string
	Location: string
	OfficialEventName: string
	EventDate: number
	EventName: string
	EventFormat: string
	Session1: string
	Session1Date: null
	Session1DateUtc: number
	Session2: string
	Session2Date: null
	Session2DateUtc: number
	Session3: string
	Session3Date: null
	Session3DateUtc: number
	Session4: string
	Session4Date: null
	Session4DateUtc: number
	Session5: string
	Session5Date: null
	Session5DateUtc: number
	F1ApiSupport: boolean
}